+++ projects/fusion-ui/components/link/v4/link.component.entities.ts
@@ -1,0 +1,3 @@
+export type LinkTarget = '_blank' | '_self' | '_parent' | '_top';
+export type LinkColor = 'default' | 'primary';
+export type LinkVariant = 'button' | 'subtitle1' | 'subtitle2' | 'body1' | 'body2' | 'caption' | 'chip-label';